// app/requests/_personal/getPlaceForBid/getPlaceForBidSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetPlaceForBidSuccess {
  data: {
    id: number;
    name: string;
    latitude: string;
    longitude: string;
    address_kladr: string;
    logo: null | string;
    region: {
      id: number;
      name: string;
      [k: string]: unknown;
    };
    brand: {
      id: number;
      name: string;
      logo: null | string;
      description: string;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}